+++ src/dummy_thumbnails/contrib/thumbnailers/sorl_thumbnail/engines/dummy_engine.py
@@ -24,7 +24,7 @@ except ImportError:
 __title__ = 'dummy_thumbnails.contrib.thumbnailers.sorl_thumbnail.' \
             'engines.dummy_engine'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016 Artur Barseghyan'
+__copyright__ = '2016-2017 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'DummyThumbnailsEngine',