+++ src/dummy_thumbnails/helpers.py
@@ -10,7 +10,7 @@ from django.conf import settings
 
 __title__ = 'dummy_thumbnails.helpers'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016 Artur Barseghyan'
+__copyright__ = '2016-2017 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('prepare_dirs_and_symlinks',)