+++ src/dummy_thumbnails/management/commands/dummy_thumbnails_symlink_dummy_images.py
@@ -1,5 +1,5 @@
 """
-Base tests module.
+Management command for creating symlinks.
 
 - Command: Management command. Prepare the dirs and create symlinks for
   testing purposes.
@@ -12,7 +12,7 @@ from ...helpers import prepare_dirs_and_symlinks
 __title__ = 'dummy_thumbnails.management.commands.' \
             'dummy_thumbnails_symlink_dummy_images'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016 Artur Barseghyan'
+__copyright__ = '2016-2017 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('Command',)